feat: reprise de l'ID si existant :sparkles:
